Next stage of refactoring
