+++ Sources/VernissageServer/Services/CollectionsService.swift
@@ -49,12 +49,12 @@ protocol CollectionsServiceType: Sendable {
 }
 
 final class CollectionsService: CollectionsServiceType {
-    private struct FeaturedCollectionData {
-        let statusIds: Set<String>
-        let statusNotes: [String: NoteDto]
-    }
 
     func synchronizeFeaturedCollection(for userId: Int64, on context: ExecutionContext) async throws {
+        let activityPubDownloadCollectionService = context.services.activityPubDownloadCollectionService
+        let activityPubDownloadStatusService = context.services.activityPubDownloadStatusService
+        let statusesService = context.services.statusesService
+
         guard let user = try await User.query(on: context.db)
             .filter(\.$id == userId)
             .first() else {
@@ -77,27 +77,8 @@ final class CollectionsService: CollectionsServiceType {
             return
         }
 
-        let usersService = context.services.usersService
-        guard let defaultSystemUser = try await usersService.getDefaultSystemUser(on: context.db) else {
-            throw ActivityPubError.missingInstanceAdminAccount
-        }
-
-        guard let privateKey = defaultSystemUser.privateKey else {
-            throw ActivityPubError.missingInstanceAdminPrivateKey
-        }
-
-        let activityPubClient = ActivityPubClient(privatePemKey: privateKey,
-                                                  userAgent: Constants.userAgent,
-                                                  host: featuredUrl.host)
-
-        let featuredCollectionData = try await self.downloadFeaturedCollectionData(featuredUrl: featuredUrl,
-                                                                                   activityPubClient: activityPubClient,
-                                                                                   activityPubProfile: defaultSystemUser.activityPubProfile,
-                                                                                   logger: context.logger)
-
+        let featuredCollectionData = try await activityPubDownloadCollectionService.downloadFeaturedCollection(featuredUrl: featuredUrl, on: context)
         let featuredStatusIds = featuredCollectionData.statusIds
-        let statusesService = context.services.statusesService
-        let activityPubDownloadService = context.services.activityPubDownloadService
 
         for featuredStatusId in featuredStatusIds {
             var status = try await statusesService.get(activityPubId: featuredStatusId, on: context.db)
@@ -121,7 +102,7 @@ final class CollectionsService: CollectionsServiceType {
 
             // When we don't have status in collection od creating failed then try to download status from remote server.
             if status == nil {
-                status = try? await activityPubDownloadService.downloadStatus(activityPubId: featuredStatusId, on: context)
+                status = try? await activityPubDownloadStatusService.download(activityPubId: featuredStatusId, on: context)
             }
 
             guard let status, status.$user.id == userId else {
@@ -145,68 +126,13 @@ final class CollectionsService: CollectionsServiceType {
     }
 
     func sendAddToFeatured(for statusId: Int64, on context: ExecutionContext) async throws {
-        try await self.sendFeaturedChange(for: statusId, type: .add, on: context)
+        let activityPubOutgoingCollectionService = context.services.activityPubOutgoingCollectionService
+        try await activityPubOutgoingCollectionService.sendFeaturedChange(for: statusId, type: .add, on: context)
     }
 
     func sendRemoveFromFeatured(for statusId: Int64, on context: ExecutionContext) async throws {
-        try await self.sendFeaturedChange(for: statusId, type: .remove, on: context)
-    }
-
-    private func sendFeaturedChange(for statusId: Int64, type: ActivityTypeDto, on context: ExecutionContext) async throws {
-        let statusesService = context.services.statusesService
-        let followsService = context.services.followsService
-        let snowflakeService = context.services.snowflakeService
-
-        guard let status = try await statusesService.get(id: statusId, on: context.db) else {
-            context.logger.warning("Cannot send '\(type.rawValue)' for featured collection. Status not found (id: \(statusId)).")
-            return
-        }
-
-        guard status.user.isLocal else {
-            context.logger.info("Skipping '\(type.rawValue)' for non-local status: '\(status.stringId() ?? "")'.")
-            return
-        }
-
-        guard let privateKey = status.user.privateKey else {
-            context.logger.warning("Cannot send '\(type.rawValue)' for featured collection. Missing private key for user '\(status.user.userName)'.")
-            return
-        }
-
-        let followersInboxes = try await followsService.getFollowersOfSharedInboxes(followersOf: status.$user.id, on: context)
-        let targetCollection = status.user.featured ?? "\(status.user.activityPubProfile)/featured"
-
-        for inbox in followersInboxes {
-            guard let inboxUrl = URL(string: inbox) else {
-                context.logger.warning("Skipping '\(type.rawValue)' for featured collection. Invalid inbox URL '\(inbox)'.")
-                continue
-            }
-
-            let activityPubClient = ActivityPubClient(privatePemKey: privateKey,
-                                                      userAgent: Constants.userAgent,
-                                                      host: inboxUrl.host)
-            let requestId = snowflakeService.generate()
-
-            do {
-                switch type {
-                case .add:
-                    try await activityPubClient.addToFeatured(objectId: status.activityPubId,
-                                                              actorId: status.user.activityPubProfile,
-                                                              targetId: targetCollection,
-                                                              on: inboxUrl,
-                                                              withId: requestId)
-                case .remove:
-                    try await activityPubClient.removeFromFeatured(objectId: status.activityPubId,
-                                                                   actorId: status.user.activityPubProfile,
-                                                                   targetId: targetCollection,
-                                                                   on: inboxUrl,
-                                                                   withId: requestId)
-                default:
-                    break
-                }
-            } catch {
-                context.logger.warning("Sending '\(type.rawValue)' to inbox failed (inbox: \(inboxUrl.absoluteString), status: \(status.activityPubId)). Error: \(error).")
-            }
-        }
+        let activityPubOutgoingCollectionService = context.services.activityPubOutgoingCollectionService
+        try await activityPubOutgoingCollectionService.sendFeaturedChange(for: statusId, type: .remove, on: context)
     }
 
     private func clearPinnedStatuses(for userId: Int64, on database: Database) async throws {
@@ -220,72 +146,4 @@ final class CollectionsService: CollectionsServiceType {
             try await pinnedStatus.save(on: database)
         }
     }
-
-    private func downloadFeaturedCollectionData(featuredUrl: URL,
-                                                activityPubClient: ActivityPubClient,
-                                                activityPubProfile: String,
-                                                logger: Logger) async throws -> FeaturedCollectionData {
-        var featuredStatusIds = Set<String>()
-        var featuredStatusNotes: [String: NoteDto] = [:]
-        var visitedPageUrls = Set<String>()
-        var nextPageUrl: URL? = featuredUrl
-        var firstPage = true
-
-        while let currentPageUrl = nextPageUrl {
-            // We have to prevent loops (when remote will return same url in next property).
-            let pageKey = currentPageUrl.absoluteString
-            if visitedPageUrls.contains(pageKey) {
-                logger.warning("Featured collection pagination loop detected for URL: '\(pageKey)'.")
-                break
-            }
-
-            // Download ordered collection.
-            visitedPageUrls.insert(pageKey)
-            let collectionDto = try await activityPubClient.featuredCollection(url: currentPageUrl, activityPubProfile: activityPubProfile)
-
-            var orderedObjects: [ObjectDto] = []
-            var firstUrlString: String?
-            var nextUrlString: String?
-
-            // Add objects to the private variable.
-            switch collectionDto {
-            case .orderedCollection(let orderedCollection):
-                orderedObjects = orderedCollection.orderedItems?.objects() ?? []
-                firstUrlString = orderedCollection.first
-            case .orderedCollectionPage(let orderedCollectionPage):
-                orderedObjects = orderedCollectionPage.orderedItems.objects()
-                nextUrlString = orderedCollectionPage.next
-            }
-
-            // Iterate via objects and add to id's or entities arrays.
-            for orderedObject in orderedObjects {
-                featuredStatusIds.insert(orderedObject.id)
-                if let noteDto = orderedObject.object as? NoteDto {
-                    featuredStatusNotes[orderedObject.id] = noteDto
-                }
-            }
-
-            // Calculate url to get next data portion.
-            if firstPage, let firstUrlString {
-                nextPageUrl = self.resolveCollectionPageUrl(firstUrlString, relativeTo: currentPageUrl)
-            } else if let nextUrlString {
-                nextPageUrl = self.resolveCollectionPageUrl(nextUrlString, relativeTo: currentPageUrl)
-            } else {
-                nextPageUrl = nil
-            }
-
-            firstPage = false
-        }
-
-        return FeaturedCollectionData(statusIds: featuredStatusIds, statusNotes: featuredStatusNotes)
-    }
-
-    private func resolveCollectionPageUrl(_ value: String, relativeTo baseUrl: URL) -> URL? {
-        if let url = URL(string: value), url.scheme != nil {
-            return url
-        }
-
-        return URL(string: value, relativeTo: baseUrl)?.absoluteURL
-    }
-
 }
